+++ apps/ledger-live-desktop/tests/page/accounts.page.ts
@@ -3,14 +3,14 @@ import { step } from "tests/misc/reporters/step";
 import { AppPage } from "tests/page/abstractClasses";
 
 export class AccountsPage extends AppPage {
-  private addAccountButton = this.page.locator("data-test-id=accounts-add-account-button");
+  private addAccountButton = this.page.getByTestId("accounts-add-account-button");
   private accountComponent = (accountName: string) =>
-    this.page.locator(`data-test-id=account-component-${accountName}`);
+    this.page.getByTestId(`account-component-${accountName}`);
   private firstAccount = this.page.locator(".accounts-account-row-item").locator("div").first();
   // Accounts context menu
-  private contextMenuEdit = this.page.locator("data-test-id=accounts-context-menu-edit");
-  private settingsDeleteButton = this.page.locator("data-test-id=account-settings-delete-button");
-  private settingsConfirmButton = this.page.locator("data-test-id=modal-confirm-button");
+  private contextMenuEdit = this.page.getByTestId("accounts-context-menu-edit");
+  private settingsDeleteButton = this.page.getByTestId("account-settings-delete-button");
+  private settingsConfirmButton = this.page.getByTestId("modal-confirm-button");
   private accountListNumber = this.page.locator(`[data-test-id^="account-component-"]`);
 
   async openAddAccountModal() {