+++ apps/ledger-live-mobile/src/const/navigation.ts
@@ -509,6 +509,13 @@ export enum ScreenName {
 
   WalletSyncActivationSettings = "WalletSyncActivationSettings",
   MockedAddAssetButton = "MockedAddAssetButton",
+  GenericLandingPage = "GenericLandingPage",
+
+  // Web3Hub
+  Web3HubMain = "Web3HubMain",
+  Web3HubSearch = "Web3HubSearch",
+  Web3HubTabs = "Web3HubTabs",
+  Web3HubApp = "Web3HubApp",
 }
 
 export enum NavigatorName {
@@ -618,4 +625,8 @@ export enum NavigatorName {
 
   AnalyticsOptInPrompt = "AnalyticsOptInPrompt",
   WalletSyncActivationSettings = "WalletSyncActivationSettings",
+  LandingPages = "LandingPages",
+
+  // Web3Hub
+  Web3Hub = "Web3Hub",
 }