+++ libs/coin-modules/coin-xrp/src/logic/getBalance.ts
@@ -1,0 +1,6 @@
+import { getAccountInfo } from "../network";
+
+export async function getBalance(address: string): Promise<bigint> {
+  const accountInfo = await getAccountInfo(address);
+  return BigInt(accountInfo.account_data.Balance);
+}