+++ libs/ledgerjs/packages/errors/src/index.ts
@@ -81,6 +81,7 @@ export const NetworkDown = createCustomErrorClass("NetworkDown");
 export const NetworkError = createCustomErrorClass("NetworkError");
 export const NoAddressesFound = createCustomErrorClass("NoAddressesFound");
 export const NotEnoughBalance = createCustomErrorClass("NotEnoughBalance");
+export const NotEnoughBalanceSwap = createCustomErrorClass("NotEnoughBalanceSwap");
 export const NotEnoughBalanceToDelegate = createCustomErrorClass("NotEnoughBalanceToDelegate");
 export const NotEnoughBalanceInParentAccount = createCustomErrorClass(
   "NotEnoughBalanceInParentAccount",
@@ -156,8 +157,19 @@ export const SyncError = createCustomErrorClass("SyncError");
 export const PairingFailed = createCustomErrorClass("PairingFailed");
 export const PeerRemovedPairing = createCustomErrorClass("PeerRemovedPairing");
 export const GenuineCheckFailed = createCustomErrorClass("GenuineCheckFailed");
-export const LedgerAPI4xx = createCustomErrorClass("LedgerAPI4xx");
-export const LedgerAPI5xx = createCustomErrorClass("LedgerAPI5xx");
+type NetworkType = {
+  status: number;
+  url: string | undefined;
+  method: string;
+};
+export const LedgerAPI4xx = createCustomErrorClass<
+  NetworkType,
+  LedgerErrorConstructor<NetworkType>
+>("LedgerAPI4xx");
+export const LedgerAPI5xx = createCustomErrorClass<
+  NetworkType,
+  LedgerErrorConstructor<NetworkType>
+>("LedgerAPI5xx");
 export const FirmwareOrAppUpdateRequired = createCustomErrorClass("FirmwareOrAppUpdateRequired");
 
 // SpeedUp / Cancel EVM tx